Add [seeder]

+++ database/migrations/2023_12_04_132521_update_projects_table.php
@@ -12,12 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->unsignedBigInteger('type_id')->after('id');
+            if (!Schema::hasColumn('projects', 'type_id')) {
+                $table->unsignedBigInteger('type_id')->nullable()->after('id');
 
-            $table->foreign('type_id')
-            ->references('id')
-            ->on('types')
-            ->onDelete('set null');
+                $table->foreign('type_id')
+                    ->references('id')
+                    ->on('types')
+                    ->onDelete('set null');
+            }
         });
     }
